# src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ApiResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"data",
					"statusCode",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaymentStatus": {
				"enum": [
					"PENDING",
					"CANCELLED",
					"COMPLETED"
				],
				"type": "string"
			},
			"OrderStatus": {
				"enum": [
					"PENDING",
					"PLACED",
					"PACKED",
					"SHIPPED",
					"DELIVERED",
					"CANCELLED",
					"RETURNED"
				],
				"type": "string"
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Add Bearer prefix on jwt token"
			}
		}
	},
	"info": {
		"title": "Market Place Apis Documentation",
		"version": "1.0.0",
		"description": "Market Place Node Backend Apis and Documentation",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"email": "marketplace@gmail.com",
			"name": "marketplace",
			"url": "http://marketplace.in"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/order/place-order": {
			"post": {
				"operationId": "PlaceOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "place order",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"shippingAddress": {
										"type": "string"
									},
									"reasonMessage": {
										"type": "string"
									},
									"rpay_signature": {
										"type": "string"
									},
									"rpay_orderId": {
										"type": "string"
									},
									"rpay_paymentId": {
										"type": "string"
									},
									"orderId": {
										"type": "string"
									}
								},
								"required": [
									"shippingAddress",
									"reasonMessage",
									"rpay_signature",
									"rpay_orderId",
									"rpay_paymentId",
									"orderId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/order/buy-now": {
			"post": {
				"operationId": "BuyNowOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "buy now order",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"shippingAddress": {
										"type": "string"
									}
								},
								"required": [
									"shippingAddress"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/order/create-rpay-order": {
			"post": {
				"operationId": "GetOrderId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "generate ship rocket order id",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"couponCode": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/order/{orderId}": {
			"put": {
				"operationId": "CancelOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "cancel order",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "get order",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/order/{orderId}/{orderItemId}": {
			"put": {
				"operationId": "CancelOrderItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "cancel order item",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "orderItemId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/order/return/{orderId}": {
			"put": {
				"operationId": "ReturnOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "return order",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/order": {
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "get orders",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "textSearch",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "paymentStatus",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentStatus"
						}
					},
					{
						"in": "query",
						"name": "orderStatus",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderStatus"
						}
					},
					{
						"in": "query",
						"name": "createdAt",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
		"/order/{orderId}/invoice": {
			"get": {
				"operationId": "GetOrdersInvoice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "get order invoice",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/order/by-vendor": {
			"get": {
				"operationId": "GetOrdersByVendor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "get orders by vendor",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "textSearch",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "paymentStatus",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentStatus"
						}
					},
					{
						"in": "query",
						"name": "createdAt",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
		"/order/by-vendor/{orderId}": {
			"get": {
				"operationId": "GetOrderByVendor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "get order by vendor",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/order/by-admin": {
			"get": {
				"operationId": "GetOrdersByAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "get orders by admin",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "textSearch",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "paymentStatus",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentStatus"
						}
					},
					{
						"in": "query",
						"name": "orderStatus",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderStatus"
						}
					},
					{
						"in": "query",
						"name": "createdAt",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
		"/order/by-admin/{orderId}": {
			"get": {
				"operationId": "GetOrderByAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "get order by Admin",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/order/users/{userId}": {
			"get": {
				"operationId": "GetOrdersByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "get orders by user Id",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "textSearch",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "paymentStatus",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentStatus"
						}
					},
					{
						"in": "query",
						"name": "orderStatus",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderStatus"
						}
					},
					{
						"in": "query",
						"name": "createdAt",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}